Update the MANIFEST.MF file we generate (for beanutils) to declare the dependency on the collections classes in a manner that can be verified by applications that support the JDK's Optional Package Versioning specification (such as Tomcat 4.0 and other servlet containers implementing the 2.3 specification).  For more info, see:
http://java.sun.com/j2se/1.3/docs/guide/extensions/versioning.html

In order to support these conventions, the recommended "name" of the
extension is the fully qualified package name of the base package
implemented by this extension.  Therefore, the generated name is now
"org.apache.commons.beanutils" instead of "beanutils".  I will adjust the
other build scripts as necessary to reflect this convention.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/beanutils/trunk@128474 13f79535-47bb-0310-9956-ffa450edef68