Make PropertyUtils.getReadMethod() and PropertyUtils.getWriteMethod() smart enough to locate accessible methods in a non-public class that are declared in a public interface implemented (either directly or indirectly) by that class.  Added unit tests to validate the behavior.
An example comes from the Crimson parser that is part of JAXP/1.1:

  // NOTE: CommentNode is *not* public
  class org.apache.crimson.tree.CommentNode
   extends org.apache.crimson.tree.DataNode
   implements org.w3c.dom.Comment;

  public interface org.w3c.dom.Comment
   extends org.w3c.dom.CharacterData;

  public interface org.w3c.dom.CharacterData
   extends org.w3c.dom.Node;

  public interface org.w3c.dom.Node;

The getNodeName() method is defined in the Node interface, but was not
accessible without this patch, because previously PropertyUtils only
checked for accessible methods in directly implemented interfaces.  Now,
it follows the interface hierarchy and correctly discovers the callable
getNodeName() method of the CommentNode class.

This bug was reported against the Struts version of PropertyUtils, along
with a patch to implement the nesting.

PR: Bugzilla #1200
Submitted by:	rashiddesai@hotmail.com


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/beanutils/trunk@128476 13f79535-47bb-0310-9956-ffa450edef68