Inspired by previous discussions on generalizing ConvertUtils, this patch makes the standard Converters used (for all the primitive types) completely pluggable, so there is no longer any concept of "custom converters".  Now, you can replace how String->int conversions (for example) are done in your application.
Besides the types previously supported by ConvertUtils (all of the Java
primitive types and their wrappers), ConvertUtils also includes standard
Converter implementations for the following Java classes:
- java.math.BigDecimal
- java.math.BigInteger
- java.sql.Date
- java.sql.Time
- java.sql Timestamp
that, in their default versions, simply use the default print format described
in the JDK JavaDocs.

One customization you might want to make is to have the standard Converters
throw ConversionException instead of returning a default value (which is the
standard behavior for backwards compatibility).  To accomplish this, the
standard converter implementations for primitives support two constructors --
one that takes a default value, and one that doesn't (and therefore will
return an exception on conversion errors).  So, if you wanted to switch to
exceptions for integer conversion errors, you would say:
  ConvertUtils.register(new IntegerConverter(), Integer.TYPE);
  ConvertUtils.register(new IntegerConverter(), Integer.class);
to cover both the wrapper class and the primitive type.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/beanutils/trunk@128547 13f79535-47bb-0310-9956-ffa450edef68