Add new standard converters (and register them) for String to primitive array conversions (for example, String --> int[]), as well as String --> String[].  The syntax accepted by these converters is similar to that accepted by a Java compiler doing array initializers, with the following adjustments:
* Only literal values of the appropriate type are recognized
  (i.e. no expressions)

* Leading and trailing "{" and "}" are allowed, but optional

* Commas between elements are considered to be whitespace,
  so that they are optional.

Thus, you can initialize a three-element integer array with values
1, 2, and 3 with any of the following String values.

  "{ 1, 2, 3}"

  "1, 2, 3"

  "{ 1 2 3 }"

  "1 2 3"

Quoted strings embedded in the initial String (delimited with either single
or double quotes) are recognized by the rules of the java.io.StreamTokenizer
class.  This is particularly useful in the String --> String[] converter,
where you can initialize array elements to a String that includes a comma
or embedded quote character.

Added some negative and positive tests for int[] and String[] conversions;
more would be useful.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/beanutils/trunk@128578 13f79535-47bb-0310-9956-ffa450edef68