Add test cases to demonstrate that a property ("dupProperty") with both subscript-based and array-based getters and setters can be utilized, including via BeanUtils.copyProperties().
BeanUtils.populate() will not be modified to deal with this, because it is
designed to convert an array of Strings into a single String if that is the
destination data type (which is what JavaBeans introspection will report in
this case).  For general purpose copying of bean properties with conversion,
BeanUtils.copyProperties() should be used instead.

PR: Bugzilla #9868
Submitted by:	Jerome Jacobsen <jerome.jacobsen at gentootech.com>


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/beanutils/trunk@128589 13f79535-47bb-0310-9956-ffa450edef68