Minor patch to refactor the implementation code a touch.
Firstly there's a patch so that this class works on HsqlDb which does not support ResultSetMetaData.getColumnClassName();

Also new Factory and Strategy methods have been added to make it easier to extend this class to, use different DynaProperty implementations and to use different class loading strategies.

Finally the class loading of column types will use the ClassLoader used to load this library, rather than the system class loader.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/beanutils/trunk@128612 13f79535-47bb-0310-9956-ffa450edef68