RowSetDynaClass is a way to represent the results of an SQL "SELECT" statement as a "disconnected row set" (in JDBC terminology).  Its function is to copy the results of a query (represented as a JDBC ResultSet or RowSet object that is passed to its constructor) into an in-memory List of the matching rows. Once the RowSetDynaClass instance has been created, the ResultSet/RowSet instance (and its corresponding Statement and Connection) can be closed and/or returned to the connection pool (if any).
In this way, RowSetDynaClass differs from ResultSetDynaClass, where the
DynaBean "instances" returned by the Iterator are really just a proxy for the
values represented in the current row of the ResultSet.

The typical usage pattern for RowSetDynaClass is in the class comments of
the Javadoc.

RowSetDynaClass instances will be Serializable, as long as all of the column
values are, which makes it a very convenient mechanism for transmitting
result sets to a remote Java-based client (such as an applet or a web
services client).


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/beanutils/trunk@128661 13f79535-47bb-0310-9956-ffa450edef68