Enhance BeanUtils.copyProperty() to deal with about 80% of the use cases for copying indexed, mapped, and nested properties with type conversions.  The remaining restrictions are documented in the Javadocs for this method.
This is a partial response to Bugzilla #16525, which documents some
restrictions in the functionality of copyProperty() that leads people
to try setProperty() instead -- which they should not do.

It is much more conservative than the proposed patch, which (as the reporter
acknowledges) is more appropriate in a minor update (1.7) versus a bugfix
patch (1.6.1) which is currently contemplated.

PR: Bugzilla #16525
Submitted by:	Tim Vernum <tpv at spamcop.net>


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/beanutils/trunk@128683 13f79535-47bb-0310-9956-ffa450edef68