Add a unit test to verify the current JDK introspection of a bean with the following method signatures:
  public boolean isFoo();
  public boolean getFoo();
  public void setFoo(String foo);

According to Section 8.3.1 of the JavaBeans Specification, this
will get recognized as a read-only boolean property using "isFoo"
as the getter method.  Therefore, all of the code in commons-beanutils
should make this same assumption as well.

One could argue that it would make more sense to not recognize "foo"
as a property at all, since the types on the getter and setter methods
are different.  However, that's not the way the JavaBeans spec is worded,
and not the way that the JDK's introspection logic works either.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/beanutils/trunk@128721 13f79535-47bb-0310-9956-ffa450edef68