bugzilla #23815.
Create new methods getPropertyOfMapBean and setPropertyOfMapBean that the existing
setNestedProperty and getNestedProperty methods now call when they discover the
bean they are accessing implements Map. This makes it much easier for users to
subclass and customise this behaviour of PropertyUtilsBean, eg in order to
restore pre-1.5 behaviour.

This patch also causes an exception to be thrown when the propertyName passed to
getPropertyOfMapBean or setPropertyOfMapBean has MAPPED_DELIM or INDEXED_DELIM
chars in it. This never worked as expected before (the whole string was treated
literally as the propertyName), so throwing an exception here should not break
any existing code. It should be of help to future developers who make this
mistake though...



git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/beanutils/trunk@178929 13f79535-47bb-0310-9956-ffa450edef68