Generified ConvertUtilsBean.
Note that the current version of the convert() methods
is not really satisfying because the generic Object type is returned.
However, we cannot return an object of the passed in target class because
the methods' behavior is to return the passed in object if no suitable
converter is found. Maybe we can add a new method which enforces the conversion
or throws an exception; this method could be type-safe.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/beanutils/branches/java5@1535854 13f79535-47bb-0310-9956-ffa450edef68