Generified LazyDynaList.
This commit also fixes [BEANUTILS-447]; the toArray() method was reworked.
Again the generification is not ideal. It would have been better to use a type
parameter (like LazyDynaList<E>); however, the current implementation as a sub
class of ArrayList and the fact that the element type can change dynamically
makes this pretty hard.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/beanutils/branches/java5@1536029 13f79535-47bb-0310-9956-ffa450edef68