Added new abstract base class for DynaBeanMapDecorators.
With DynaBeanMapDecorator there is the problem that the correct generic type
parameters cannot be used because this would break backwards compatibility.
Therefore, a new abstract base class was added which accepts a type parameter
for the decorated map's key type. This makes it very easy to create multiple
concrete decorator classes with different key types. So DynaBeanMapDecorator
can be kept as it is, and a new decorator class using String as key type can
be added easily. (We can think about deprecating DynaBeanMapDecorator in
favor of the new decorator with correct generic types.)

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/beanutils/branches/java5@1536075 13f79535-47bb-0310-9956-ffa450edef68