Generified locale converters.
The locale converters now provide a convert method with a type parameter
defining the target conversion type. This commit also fixes
[BEANUTILS-448] and [BEANUTILS-449] because the conversion result is now
explicitly checked whether it is compatible with the desired target type.
This is a change in the behavior of the locale converters; therefore some tests
had to be adapted.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/beanutils/branches/java5@1537267 13f79535-47bb-0310-9956-ffa450edef68